Centralize Synonyms Check for Tags
