+++ src/__tests__/mentors.json-test.js
@@ -3,6 +3,7 @@ import Ajv from 'ajv';
 import { get as getPath } from 'object-path';
 import countries from 'svg-country-flags/countries.json';
 import _ from 'lodash';
+import checkSynonyms from '../checkSynonymsTags';
 
 expect.extend({
   toBeValid(isValid, errorMessage) {
@@ -26,28 +27,15 @@ const validateSecuredUrl = function(schema, uri) {
   return uri.indexOf('https://') === 0;
 };
 
-const synonymsTags = {
-  '(node|node.js)': 'nodejs',
-  '(react|React.js)': 'reactjs',
-  vue: 'vuejs',
-  'react-native': 'react native',
-  csharp: 'c#',
-  'front end': 'frontend',
-  expressjs: 'express',
-  'full stack': 'fullstack',
-};
-
 const validateSynonymsTags = function(schema, tag) {
   let isValid = true;
   let message = '';
-  Object.keys(synonymsTags).forEach(synonym => {
-    if (new RegExp(`^${synonym}$`, 'i').exec(tag)) {
-      message = `should NOT use "${tag}", should use the conventional name: "${
-        synonymsTags[synonym]
-      }"`;
-      isValid = false;
-    }
-  });
+  const synonymError = checkSynonyms(tag);
+
+  if (synonymError) {
+    message = synonymError;
+    isValid = false;
+  }
 
   validateSynonymsTags.errors = [
     { keyword: 'synonymsTags', message, params: { keyword: 'synonymsTags' } },
